Crate cranelift_codegen_shared

This library contains code that is common to both the cranelift-codegen and cranelift-codegen-meta libraries.

Modules

condcodes
Condition codes for the Cranelift code generator.

constant_hash
Build support for precomputed constant hash tables.

constants
This module contains constants that are shared between the codegen and the meta crate, so they are kept in sync.

isa
Shared ISA-specific definitions.

Constants

VERSION
Version number of this crate.