Module cranelift_codegen_shared::condcodes

Condition codes for the Cranelift code generator.

A condition code here is an enumerated type that determined how to compare two numbers. There are different rules for comparing integers and floating point numbers, so they use different condition codes.

Enums

FloatCC
Condition code for comparing floating point numbers.

IntCC
Condition code for comparing integers.

Traits

CondCode
Common traits of condition codes.