Function cranelift_codegen_shared::constant_hash::generate_table

pub fn generate_table<'cont, T, I: Iterator<Item = &'cont T>, H: Fn(&T) -> usize>(
    items: I, 
    num_items: usize, 
    hash_function: H
) -> Vec<Option<&'cont T>>

Compute an open addressed, quadratically probed hash table containing items. The returned table is a list containing the elements of the iterable items and None in unused slots.