cranelift_codegen_shared

Module constants

This module contains constants that are shared between the codegen and the meta crate, so they are kept in sync.

Constants

DYNAMIC_VECTOR_BASE

Start of the dynamic vector types.

LANE_BASE

Start of the lane types.

REFERENCE_BASE

Base for reference types.

VECTOR_BASE

Start of the 2-lane vector types.