Crate cranelift_codegen_shared

This library contains code that is common to both the cranelift-codegen and cranelift-codegen-meta libraries.

Modules

constant_hash

This module provides a primitive hash function.

constants

This module contains constants that are shared between the codegen and the meta crate, so they are kept in sync.

Constants

VERSION

Version number of this crate.